fix SetContextId integer truncation

AI[89%] Human Fixed[0%] Human[11%] AI Adopted[100%]

+++ services/accountmgr/src/account_iam/account_iam_callback.cpp
@@ -58,7 +58,7 @@ using UserAuthClient = UserIam::UserAuth::UserAuthClient;
 
 const std::vector<uint8_t> TEMP_PIN = {50, 48, 50, 52, 48, 56};
 
-void AuthCallbackDeathRecipient::SetContextId(uint16_t contextId)
+void AuthCallbackDeathRecipient::SetContextId(uint64_t contextId)
 {
     contextId_ = contextId;
 }
